+++ FreeRTOS/Source/portable/GCC/STR75x/portISR.c
@@ -78,9 +78,9 @@
 #include "task.h"
 
 /* Constants required to handle critical sections. */
-#define portNO_CRITICAL_NESTING		( ( unsigned long ) 0 )
+#define portNO_CRITICAL_NESTING		( ( uint32_t ) 0 )
 
-volatile unsigned long ulCriticalNesting = 9999UL;
+volatile uint32_t ulCriticalNesting = 9999UL;
 
 /*-----------------------------------------------------------*/
